Power on/off changes + misc fixes
Corrected the power on/off circuit (connected the top of the switch to the right net) and bumped up the resistor value so that it'll take ~6 seconds of the power button being held down to power off.
The circuit should now match Figure 5 from [this](http://www.mosaic-industries.com/embedded-systems/microcontroller-projects/electronic-circuits/push-button-switch-turn-on/latching-toggle-power-switch).

This fixes #1.

Additionally, libraries used in the V0 files should match the rearranged folders.