Add an experimental card detect input to the ADC button circuit
If this is problematic we can just remove/not populate the resistor for CARD_DETECT (R54). Just adding an input to the ADC and not the full diode + digital input line too because we can just poll the ADC when we have SD card errors to make sure we've got a card in place (I'd rather not add another one of those diodes).