Bridge IO0 and IO2 so that we can flash the board and use an SD Card
From the message on sheet 1:
IO2 is a strapping pin that allows the esp32 to enter download mode (iirc) - it's required to be pulled low when flashing the esp. Unforuntately, we have to use IO2 for the one usable SD/MMC interface on the esp32 (the other uses the same pins as flash).

IO0 is also a strapping pin: it's connected to RTS on the CP2102N and is pulled low when we're flashing the board. So, if we tie IO0 to IO2, IO2 will be pulled down while flashing. This solves our problem!

Unforunately this means that we can't use IO0 for anything else since it remains tied to IO2 (anything we do to IO0 will affect the SD/MMC peripheral). We could use a fancy circuit that pulls IO2 low only when the chip is reset (for a second or so) or do some transistor magic (only pull IO2 low when multiple other pins are low or something), but since we're only using IO0 for an LED, I figure it's not worth the trouble.

We'll move the LED's net to something else and connect IO0 and IO2.

Ended up putting the green and blue LEDs on the same pin. Since the voltage drops are different, hopefully we can get slightly different colors by PWMing at different frequencies.