VS1053 changes
Switch back to bridging all the GPIO pins (the datasheet's sample connection diagram has it wired this way, so it'll hopefully not be an issue) and tie xRESET to V3P3 (pull-up) so that we can actually boot.
The Adafruit VS1053 breakout schematic mentions that they tie xRESET to low (pull-down) to avoid transients (clicks) in the headphone outputs (as per [this](http://www.vlsi.fi/uploads/media/vs10XXan.pdf)) - making the MCU enable the VS1053 some time after system power up apparently helps mitigate this. Adafruit goes a step further and says that not doing this can "damage the earphones", but since we're pretty heavily pin starved we'll have to risk it.

At this point we seem to match Adafruit's VS1053 breakout in most of the relevant ways. The main (relevant) differences are that we don't differentiate between AGND and GND (or have ferrite beads between them), don't have pull-ups on our CS pins (though this is probably okay since the ESP32 [supposedly doesn't need external pull-ups on its SPI CS pins](https://esp32.com/viewtopic.php?t=3650)), don't have the 2nd set of capacitors/resistors on our right/left output lines (ESD?), and don't have what looks to be a bypass cap and a pull-up (??) resistor on xRESET (peculiar since there's a pull-down resistor on the other side of the level shifter for this pin).

Anyways, we should be in decent shape and we've got the header pins as an escape hatch in case things don't work out.